+++ R/namespace.R
@@ -1,4 +1,5 @@
-#' @import coda err checkr
+#' @import err checkr
+#' @import coda
 #' @importFrom graphics plot
 #' @importFrom stats coef terms
 NULL