+++ R/rhat.R
@@ -8,7 +8,7 @@
 #' @param by A string indicating whether to determine by
 #' "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the values as a
-#' tibble versus a named list.
+#' data frame versus a named list.
 #' @return The rhat value(s).
 #' @references
 #' Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
@@ -49,8 +49,8 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, ...) {
       return(NA_real_)
     }
     if(by == "term")
-      return(tibble::tibble(term = terms(x), rhat = NA_real_))
-    return(tibble::tibble(parameter = "parameter", rhat = NA_real_))
+      return(tibble(term = terms(x), rhat = NA_real_))
+    return(tibble(parameter = "parameter", rhat = NA_real_))
   }
 
   x <- split_chains(x)
@@ -61,7 +61,7 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, ...) {
     return(max(x))
   }
   if(by != "term")
-    return(tibble::tibble(parameter = "parameter", rhat = max(x)))
+    return(tibble(parameter = "parameter", rhat = max(x)))
   x <- estimates(as.mcmcarray(x), as_df = TRUE)
   colnames(x) <- c("term", "rhat")
   x
@@ -79,7 +79,7 @@ rhat.mcmcr <- function(x, by = "all", as_df = FALSE, ...) {
   x <- Map(x, parameters, f = function(x, p) {parameters(x[[1]]) <- p; x})
   x <- do.call(rbind, x)
   if(by == "all")
-    return(tibble::tibble(all = "all", rhat = max(x$rhat)))
+    return(tibble(all = "all", rhat = max(x$rhat)))
   x
 }