+++ R/utils.R
@@ -70,7 +70,7 @@ pvalue <- function(x) {
 #' @param x An MCMC object.
 #' @param by A string indicating whether to determine by "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the results as a
-#' tibble or list.
+#' data frame or list.
 #' @export
 #' @examples
 #' ess(mcmcr_example)