+++ NEWS.md
@@ -1,4 +1,4 @@
-- from imports to suggests: graphics, coef
+- from imports to suggests: graphics, coef, tibble
 - replace internal error() with err::err()
 
 # mcmcr 0.0.2